Fix price attributes field type
The field type is changed to float to prevent issues in calculations.

+++ app/models/basket.rb
@@ -5,6 +5,6 @@ class Basket < ApplicationRecord
   has_many :items, through: :basket_items
   belongs_to :user
 
-  attribute :price, :integer, default: 0
-  attribute :final_price, :integer, default: 0
+  attribute :price, :float, default: 0.0
+  attribute :final_price, :float, default: 0.0
 end
